+++ src/app/Balloon.App.Convert/App/Http.php
@@ -14,19 +14,19 @@ namespace Balloon\App\Convert\App;
 
 use Balloon\App\AppInterface;
 use Balloon\App\Convert\Api\v1\Convert;
-use Micro\Http\Router;
-use Micro\Http\Router\Route;
 use Balloon\Filesystem\Node\AttributeDecorator;
 use Balloon\Server;
+use Micro\Http\Router;
+use Micro\Http\Router\Route;
 
 class Http implements AppInterface
 {
     /**
      * Constructor.
      *
-     * @param Router $router
+     * @param Router             $router
      * @param AttributeDecorator $decorator
-     * @param Server $server
+     * @param Server             $server
      */
     public function __construct(Router $router, AttributeDecorator $decorator, Server $server)
     {
@@ -36,16 +36,15 @@ class Http implements AppInterface
             ->prependRoute(new Route('/api/v1/file/convert', Convert::class))
             ->prependRoute(new Route('/api/v1/file/{id:#([0-9a-z]{24})#}/convert', Convert::class));
 
-        $decorator->addDecorator('master', function($node, $attributes) use($fs, $decorator) {
+        $decorator->addDecorator('master', function ($node, $attributes) use ($fs, $decorator) {
             $master = $node->getAppAttribute('Balloon\\App\\Convert', 'master');
-
-            if($master === null) {
+            if (null === $master) {
                 return null;
             }
 
             try {
-                return $decorator->decorate($fs->getNodeById($master), $attributes);
-            } catch(\Exception $e) {
+                return $decorator->decorate($fs->findNodeById($master), $attributes['attributes']);
+            } catch (\Exception $e) {
                 return null;
             }
         });