+++ src/lib/Filesystem.php
@@ -502,7 +502,7 @@ class Filesystem
     public function findNodesByFilterUser(int $deleted, array $filter): Generator
     {
         if ($this->user instanceof User) {
-            $this->user->findNewShares();
+//            $this->user->findNewShares();
         }
 
         $shares = $this->user->getShares();