+++ src/lib/Filesystem/Node/File.php
@@ -176,7 +176,6 @@ class File extends AbstractNode implements FileInterface
                 'changed' => $this->changed,
                 'deleted' => $this->deleted,
                 'meta' => $this->meta,
-                'app_attributes' => $this->app_attributes,
             ], NodeInterface::CONFLICT_NOACTION, true);
         }